Treat relative root nodes as root nodes
When a tree is cut in more than one page (because is long), only the
first page has the real root node of this tree. But for the others pages
we still want to display the nodes, so treat the less deep node as the
"current relative root" node.
By the way, I think we could remove the "node.r === True" stuff and replace
it with something like "node.l === 0". And thus make the json lighter.